+++ src/pebbledoc/cli_parser.py
@@ -154,6 +154,19 @@ def _build_parser() -> argparse.ArgumentParser:
         default=None,
     )
     rendering_group.add_argument(
+        "--main-docstring",
+        help=(
+            "position for the main docstring of the package:\n"
+            "- default: place the package docstring in its dedicated section "
+            "(default)\n"
+            "- pre: places the docstring before the table of contents\n"
+            "- post: places the docstring after the table of contents\n"
+            "- omit: omits the main docstring entirely"
+        ),
+        choices=["default", "pre", "post", "omit"],
+        default=None,
+    )
+    rendering_group.add_argument(
         "--title",
         help="set the title for the document (i.e. its main header)",
         metavar="",
@@ -161,6 +174,11 @@ def _build_parser() -> argparse.ArgumentParser:
     )
 
     formatting_group = parser.add_argument_group(title="formatting")
+    formatting_group.add_argument(
+        "--no-generic-intro",
+        help="omit the generic introduction after the main header",
+        action="store_true",
+    )
     formatting_group.add_argument(
         "--no-module-docstring",
         help="omit module-level docstrings for submodules and sub-packages",